Bring Winforms back

+++ UTDR-Music-Randomizer/UTDR-Music-Randomizer/UTDRRandomizer.cs
@@ -16,7 +16,7 @@ namespace UTDRMusicRandomizer
 
         private static string[] CreditsSongsDeltarune = new string[]
         {
-            "dontforget", "ch2_credits"
+            "dontforget", "ch2_credits", "ch4_credits"
         };
 
         private static string[] CreditsSongsUndertale = new string[]
@@ -36,9 +36,12 @@ namespace UTDRMusicRandomizer
 
         private static string[] AmbienceDeltarune = new string[]
         {
-            "alley_ambience", "audio_drone", "bird", "deep_noise", "elevator",
-            "honksong", "mus_birdnoise", "ocean", "shinkansen", "sink_noise",
-            "w", "wind_highplace", "wind"
+            "alley_ambience", "audio_drone", "bell_ambience", "bird", "board_ocean",
+            "carol_appeared", "church_lightning", "deep_noise", "elevator", "heartbeat",
+            "honksong", "knight_appears", "muffled_rain", "mus_birdnoise",
+            "mus_knightthought", "night_ambience", "noelleshouseoutside", "ocean", "raining",
+            "raining_in_church2", "shinkansen", "sink_noise", "strongwind_loop",
+            "tv_infrontof", "w", "wind_highplace", "wind"
         };
 
         private static string[] AmbienceUndertale = new string[]
@@ -50,11 +53,33 @@ namespace UTDRMusicRandomizer
             "mus_wind", "mus_zzz_c", "mus_zzz_c2"
         };
 
+        private static string[] RhythmGameSongs = new string[]
+        {
+            "battle_preview", "battle_preview_crisp", "board4_rhythm",
+            "boxing_boss_preview", "boxing_boss_preview_crisp",
+            "ch3_karaoke_example_guit_only", "ch3_karaoke_example_guitar_only_v2",
+            "ch3_karaoke_example_noguit", "ch3_karaoke_example_noguitar_v2",
+            "ch3_karaoke_example_together_ex", "ch3_karaoke_full",
+            "ch3_karaoke_no_guitar", "ch3_karaoke_preview",
+            "ch3_karaoke_preview_crisp", "ch3_tvtime_guitar",
+            "ch3-practice_song_combined", "ch3-practice_song_noguit",
+            "field_of_hopes_preview", "field_of_hopes_preview_crisp",
+            "rhythm_knockdown_combined", "rhythm_knockdown_no_guit",
+            "tenna_battle_guitar", "tenna_battle_preview", "tenna_battle_preview_crisp"
+        };
+
         private static string[] SoundsDeltarune = new string[]
         {
-            "berdly_audience", "berdly_descend", "charjoined", "cyber_battle_end",
-            "fanfare", "queen_intro", "s_neo_clip", "spamton_laugh_noise",
-            "static_placeholder", "tv_noise"
+            "berdly_audience", "berdly_descend", "board_lancer_dig", "charjoined",
+            "ch4_first_intro", "ch4_first_intro_breaking", "cyber_battle_end",
+            "fanfare", "me", "ominous_message", "ominous_stab_harsh", "ominous_stab_harsh_2",
+            "ominous_worse", "pianpian", "queen_intro", "rtenna_zoom", "s_neo_clip",
+            "sadchord2", "sinedrone_danger", "sinedrone_danger_high", "sound_battle_bg",
+            "statue_chord_basic", "statue_level1", "statue_level2", "statue_level4",
+            "statue2_level1", "statue2_level2", "statue2_level3", "statue2_level4",
+            "statue2_level5", "spamton_laugh_noise", "static_placeholder", "tinnitus",
+            "tv_noise", "tv_static_bad", "wet_tapdancing", "wet_tapdancing_failed",
+            "wet_tapdancing2"
         };
 
         private static string[] SoundsUndertale = new string[]
@@ -73,13 +98,15 @@ namespace UTDRMusicRandomizer
 
         private struct SongFile
         {
-            public SongFile(string name, bool fromBaseFolder = false)
+            public SongFile(string name, bool fromBaseFolder = false, int chapter = 0)
             {
                 Name = name;
                 FromBaseFolder = fromBaseFolder;
+                Chapter = chapter;
             }
             public string Name { get; }
             public bool FromBaseFolder { get; }
+            public int Chapter { get; }
         }
 
         public static bool IsValidPath(string path)
@@ -155,6 +182,14 @@ namespace UTDRMusicRandomizer
             return musicPath;
         }
 
+        private static string GetChapterPath(string basePath, int chapter)
+        {
+            string path = GetBasePath(basePath);
+            string platform = "windows";
+            if (IsMac(basePath)) { platform = "mac"; }
+            return path + GetSeparator(basePath) + $"chapter{chapter}_{platform}";
+        }
+
         private static string GetBackupPath(string basePath)
         {
             return GetMusicPath(basePath) + GetSeparator(basePath) + "backup";
@@ -165,12 +200,20 @@ namespace UTDRMusicRandomizer
             return Directory.Exists(GetBackupPath(basePath));
         }
 
+        private static bool IsFromBaseFolder(string name)
+        {
+            return name.StartsWith("snd_") || name.ToLower().StartsWith("audio_intronoise") ||
+                name.StartsWith("mus_undynescary");
+        }
+
         public static RandoResult Randomize(string basePath, RandoOptions options, bool makeBackup)
         {
             string musicPath = GetMusicPath(basePath);
             bool isDeltarune = IsDeltarune(basePath);
             string[] files = Directory.GetFiles(musicPath);
             IEnumerable<string> oggFiles;
+            int i;
+
             if (isDeltarune)
             {
                 oggFiles =
@@ -180,6 +223,7 @@ namespace UTDRMusicRandomizer
                     where (options.CyberBattle ? true : Path.GetFileNameWithoutExtension(f) != "cyber_battle_prelude")
                     where (options.CreditsSongs ? true : !CreditsSongsDeltarune.Contains(Path.GetFileNameWithoutExtension(f).ToLower()))
                     where (options.Ambience ? true : !AmbienceDeltarune.Contains(Path.GetFileNameWithoutExtension(f).ToLower()))
+                    where (options.RhythmGame ? true : !RhythmGameSongs.Contains(Path.GetFileNameWithoutExtension(f)))
                     where (options.SFX ? true : !SoundsDeltarune.Contains(Path.GetFileNameWithoutExtension(f).ToLower()))
                     select f;
             }
@@ -231,44 +275,80 @@ namespace UTDRMusicRandomizer
 
                         if (isDeltarune && options.SFX) //grab SFX from base folder
                         {
-                            var drFiles = Directory.GetFiles(GetBasePath(basePath));
-                            var drSounds =
-                                from f in drFiles
-                                where Path.GetExtension(f) == ".ogg"
-                                select f;
-
-                            foreach (var f in drSounds)
+                            //check if chapters are separated (newer versions of Deltarune)
+                            if (Path.Exists(GetChapterPath(basePath, 1)))
                             {
-                                temp = Path.GetFileName(f);
-                                backupFilePath = backupFolder.FullName + separator + temp;
-                                tempNames.Add(new SongFile(temp, true));
-                                if (makeBackup || !File.Exists(backupFilePath))
+                                for (i = 1; i <= 7; ++i)
                                 {
-                                    File.Delete(backupFilePath);
-                                    File.Move(f, backupFilePath);
+                                    string path = GetChapterPath(basePath, i);
+                                    if (Path.Exists(path))
+                                    {
+                                        var chFiles = Directory.GetFiles(path);
+                                        var sounds = (from f in chFiles
+                                            where Path.GetExtension(f) == ".ogg"
+                                            select f);
+
+                                        foreach (var f in sounds)
+                                        {
+                                            temp = Path.GetFileName(f);
+                                            backupFilePath = backupFolder.FullName + separator + temp;
+                                            tempNames.Add(new SongFile(temp, true, i));
+                                            if (makeBackup || !File.Exists(backupFilePath))
+                                            {
+                                                File.Delete(backupFilePath);
+                                                File.Move(f, backupFilePath);
+                                            }
+                                            else
+                                            {
+                                                File.Delete(f);
+                                            }
+                                        }
+                                    }
                                 }
-                                else
+                            }
+                            else //if not, get files from the base folder
+                            {
+                                var drFiles = Directory.GetFiles(GetBasePath(basePath));
+                                var sounds = (from f in drFiles
+                                    where Path.GetExtension(f) == ".ogg"
+                                    select f);
+
+                                foreach (var f in sounds)
                                 {
-                                    File.Delete(f);
+                                    temp = Path.GetFileName(f);
+                                    backupFilePath = backupFolder.FullName + separator + temp;
+                                    tempNames.Add(new SongFile(temp, true));
+                                    if (makeBackup || !File.Exists(backupFilePath))
+                                    {
+                                        File.Delete(backupFilePath);
+                                        File.Move(f, backupFilePath);
+                                    }
+                                    else
+                                    {
+                                        File.Delete(f);
+                                    }
                                 }
                             }
                         }
 
                         var rng = new Random();
-                        int i = 0;
+                        i = 0;
                         string tempPath;
                         foreach (var f in backupFolder.GetFiles())
                         {
                             tempPath = (musicPath + separator + f.Name).Replace($"{separator}{separator}", $"{separator}");
-                            if (oggFiles.Contains(tempPath) || (isDeltarune && options.SFX &&
-                                (f.Name.StartsWith("snd_") || f.Name.StartsWith("audio_intronoise"))))
+                            if (oggFiles.Contains(tempPath) || (isDeltarune && options.SFX && IsFromBaseFolder(f.Name)))
                             {
                                 do
                                 {
                                     i = rng.Next(tempNames.Count);
                                 } while (tempNames[i].Name == Path.GetFileName(f.Name));
 
-                                if (tempNames[i].FromBaseFolder)
+                                if (tempNames[i].Chapter > 0)
+                                {
+                                    temp = GetChapterPath(basePath, tempNames[i].Chapter) + separator + tempNames[i].Name;
+                                }
+                                else if (tempNames[i].FromBaseFolder)
                                 {
                                     temp = basePath + separator + tempNames[i].Name;
                                 }
@@ -326,21 +406,45 @@ namespace UTDRMusicRandomizer
             try
             {
                 string musicPath = GetMusicPath(basePath), backupPath = GetBackupPath(basePath);
-                bool isDeltarune = IsDeltarune(basePath);
+                bool isDeltarune = IsDeltarune(basePath), checkAllChapters;
+
                 foreach (var f in Directory.GetFiles(backupPath))
                 {
                     string destPath = musicPath, fileName = Path.GetFileName(f);
-                    if (isDeltarune && (fileName.StartsWith("snd_") || fileName.ToLower().StartsWith("audio_intronoise")))
+                    checkAllChapters = false;
+                    if (isDeltarune && IsFromBaseFolder(fileName))
                     {
                         destPath = GetBasePath(basePath);
+                        checkAllChapters = Path.Exists(GetChapterPath(basePath, 1));
                     }
-                    destPath += GetSeparator(basePath) + fileName;
 
-                    if (File.Exists(destPath))
+                    if (checkAllChapters) //copy this file to each chapter that has it
                     {
-                        File.Delete(destPath);
+                        for (int i = 1; i <= 7; ++i)
+                        {
+                            destPath = GetChapterPath(basePath, i);
+                            if (Path.Exists(destPath))
+                            {
+                                destPath += GetSeparator(basePath) + fileName;
+
+                                if (File.Exists(destPath))
+                                {
+                                    File.Delete(destPath);
+                                    File.Copy(f, destPath);
+                                }
+                            }
+                        }
                     }
-                    File.Copy(f, destPath);
+                    else //just copy it to the correct folder
+                    {
+                        destPath += GetSeparator(basePath) + fileName;
+
+                        if (File.Exists(destPath))
+                        {
+                            File.Delete(destPath);
+                        }
+                        File.Copy(f, destPath);
+                    }
                 }
                 return true;
             }
